pyproject.toml
[tool.poetry]
name = "samapi"
version = "0.1.0"
description = "Segment Anything Model (SAM) API"
authors = ["Ko Sugawara <ko.sugawara@ens-lyon.fr>"]
readme = "README.md"
packages = [{include = "samapi", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
fastapi = {extras = ["all"], version = "^0.96.0"}
torch = "^2.0.1"
segment-anything = {git = "https://github.com/facebookresearch/segment-anything.git"}
geojson = "^3.0.1"
pillow = "^9.5.0"
numpy = "^1.24.3"
scikit-image = "^0.21.0"
shapely = "^2.0.1"
torchvision = "^0.15.2"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
flake8 = "^6.0.0"
ipython = "^8.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
